+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Feature
+- Add pixel in AddProductBtn and removeProductBtn
 ## [1.8.4] - 2021-08-31
-
 
 ### Fix
 - Add view empty when user has only one product and this product has no inventory